# Optimize OpenMP lock contentions