This PR aims at optimizing some OpenMP matrix conversions and other operations that cause a lot of lock contention and long test runtimes when you execute them with many cores. 

TODO

- [x] dense -> hybrid (~6000ms)
- [x] ~~gmres initialize_1/2 (4000ms)~~
- [x] ~~gmres step_1 (7000ms)~~
- [x] ~~gmres step_2 (3000ms)~~